requirements.txt
# Python dependencies for YOLOv5 Object Detection on DIOR Dataset
numpy>=1.18.5
matplotlib>=3.2.2
opencv-python>=4.1.2
Pillow>=8.0.0
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
pandas>=1.1.4
seaborn>=0.11.0
scikit-learn>=0.24.1
ipykernel>=6.0.3
notebook>=6.4.0
gdown>=4.4.0  # For downloading files from Google Drive 